+++ .run/api.run.xml
@@ -1,0 +1,11 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="api" type="GoApplicationRunConfiguration" factoryName="Go Application">
+    <module name="bird-lg-go" />
+    <working_directory value="$PROJECT_DIR$" />
+    <kind value="PACKAGE" />
+    <filePath value="$PROJECT_DIR$/api/main.go" />
+    <package value="github.com/NULLx76/bird-lg-go/api" />
+    <directory value="$PROJECT_DIR$" />
+    <method v="2" />
+  </configuration>
+</component>
