+++ .run/frontend.run.xml
@@ -1,0 +1,14 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="frontend" type="GoApplicationRunConfiguration" factoryName="Go Application">
+    <module name="bird-lg-go" />
+    <working_directory value="$PROJECT_DIR$" />
+    <kind value="PACKAGE" />
+    <filePath value="$PROJECT_DIR$/frontend/main.go" />
+    <package value="github.com/NULLx76/bird-lg-go/frontend" />
+    <directory value="$PROJECT_DIR$" />
+    <method v="2">
+      <option name="RunConfigurationTask" enabled="true" run_configuration_name="build" run_configuration_type="js.build_tools.npm" />
+      <option name="RunConfigurationTask" enabled="true" run_configuration_name="minify" run_configuration_type="js.build_tools.npm" />
+    </method>
+  </configuration>
+</component>
