+++ frontend/css/style.pcss
@@ -4,34 +4,36 @@
 
 @tailwind utilities;
 
-h1 {
-    @apply text-4xl text-indigo-700 font-bold
-}
+@layer base {
+    h1 {
+        @apply text-4xl text-indigo-700 font-bold
+    }
 
-h2 {
-    @apply text-xl font-medium text-gray-900;
-}
+    h2 {
+        @apply text-xl font-medium text-gray-900;
+    }
 
-h3 {
-    @apply text-lg font-medium;
-}
+    h3 {
+        @apply text-lg font-medium;
+    }
 
-a {
-    @apply text-blue-700;
-}
+    a {
+        @apply text-blue-700;
+    }
 
-table {
-    @apply border-collapse border-2 border-gray-500;
-}
+    table {
+        @apply border-collapse border-2 border-gray-500;
+    }
 
-th {
-    @apply border border-gray-400 px-4 py-2 text-gray-800;
-}
+    th {
+        @apply border border-gray-400 px-4 py-2 text-gray-800;
+    }
 
-td {
-    @apply border border-gray-400 px-4 py-2 text-center;
-}
+    td {
+        @apply border border-gray-400 px-4 py-2 text-center;
+    }
 
-pre {
-    @apply p-2 m-2 border-2 border-gray-400;
+    pre {
+        @apply p-2 m-2 border-2 border-gray-400;
+    }
 }
